Add app.js file for api calls to live
